requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.5.0
et-xmlfile==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-admin-interface==0.16.1
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
et-xmlfile==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlwt==1.3.0
gunicorn==20.1.0
whitenoise==5.2.0
psycopg2==2.8.6
dj-database-url==0.5.0
